fix BB splitting
BB split moves targets from the old BB to the new tail. If there
were more than 1 out edges, we skipped some, because we used an
out_edge_iterator and removed an element in between, which is
obviously unsafe. Hence, we now get a new iterator for every
loop iteration.