dynlist: allow to specifiy EIP of iList start
some benchmarks (e.g., the quicksort examples) contain an initialization
phase that we do not want to be part of the PVF analysis. Therefore we
added a '-i' option to dynrun to tell it that all instructions until the
first occurrence of the address specified on the cmdline shall be
ignored.